chore(CRT-356): move template.Processor logic to common
